Engine: More tweaking of the zooming code. Spoken dialogue should now render properly, even when zoomed in.
git-svn-id: svn://svn.code.sf.net/p/opensludge/code@633 27a91dd4-ec73-4ff0-838b-64dfa05f697e

+++ source/Engine/backdrop.cpp
@@ -62,7 +62,7 @@ unsigned int sceneWidth, sceneHeight;
 int lightMapNumber;
 unsigned int currentBlankColour = makeColour (0, 0, 0);
 
-extern int cameraX, cameraY;
+extern double cameraX, cameraY;
 extern float cameraZoom;
 
 void nosnapshot () {
